Added some better error handling to build script

+++ invoke-docker-build.ps1
@@ -26,6 +26,17 @@ if (-not (Get-Command "dotnet" -EA SilentlyContinue))
     throw "This script requires dotnet cli for building the service"
 }
 
+if (-not (Get-Command "npm" -EA SilentlyContinue))
+{
+    throw "This script requires npm for building the UI"
+}
+
+if (-not (Get-Command "ng" -EA SilentlyContinue))
+{
+    Write-Host -ForegroundColor Magenta "Run: npm install -g @angular/cli"
+    throw "This script requires angular for building the UI"
+}
+
 if ($Version)
 {
     $Version = "$Version-"
@@ -35,22 +46,38 @@ $ImageName = "oneidentity/safeguard-devops:$Version$ImageType"
 try
 {
     Push-Location $PSScriptRoot
-    Write-Host "Cleaning up all build directories ..."
+    Write-Host -ForegroundColor Yellow "Cleaning up all build directories ..."
     (Get-ChildItem -Recurse -Filter obj -EA SilentlyContinue) | ForEach-Object { Remove-Item -Recurse -Force $_.FullName }
     (Get-ChildItem -Recurse -Filter bin -EA SilentlyContinue) | ForEach-Object { Remove-Item -Recurse -Force $_.FullName }
-    Write-Host "Building for full-size Linux distros ..."
+    Write-Host -ForegroundColor Yellow "Building for full-size Linux distros ..."
     dotnet publish -v d -r linux-x64 -c Release --self-contained --force /p:PublishSingleFile=true SafeguardDevOpsService/SafeguardDevOpsService.csproj
-    Write-Host "Building for tiny Linux distros ..."
+    if ($LASTEXITCODE -ne 0)
+    {
+        throw "dotnet publish command failed"
+    }
+    Write-Host -ForegroundColor Yellow "Building for tiny Linux distros ..."
     dotnet publish -v d -r linux-musl-x64 -c Release --self-contained --force /p:PublishSingleFile=true SafeguardDevOpsService/SafeguardDevOpsService.csproj
+    if ($LASTEXITCODE -ne 0)
+    {
+        throw "dotnet publish command failed"
+    }
 
     if (Invoke-Expression "docker images -q $ImageName")
     {
-        Write-Host "Cleaning up the old image: $ImageName ..."
+        Write-Host -ForegroundColor Yellow "Cleaning up the old image: $ImageName ..."
         & docker rmi --force "$ImageName"
+        if ($LASTEXITCODE -ne 0)
+        {
+            throw "docker command failed"
+        }
     }
 
-    Write-Host "Building a new image: $ImageName ..."
+    Write-Host -ForegroundColor Yellow "Building a new image: $ImageName ..."
     & docker build --no-cache -t "$ImageName" -f "$SafeguardDockerFile" "$PSScriptRoot"
+    if ($LASTEXITCODE -ne 0)
+    {
+        throw "docker command failed"
+    }
 }
 finally
 {
